refactor(genja-core): Split settings helpers and tests into submodules.
