+++ genja-core/src/settings.rs
@@ -60,211 +60,29 @@
 //! ## SSH Validation
 //! SSH config is validated automatically when calling `Settings::from_file`.
 //! For manual validation, use `SSHConfig::validate`.
-use crate::inventory::{Defaults, Groups, Hosts, TransformFunctionOptions};
-use crate::{ConfigLoadError, InventoryFileKind, InventoryLoadError, SshConfigError};
+use crate::inventory::TransformFunctionOptions;
+use crate::ConfigLoadError;
 use config::{Config as ConfigBuilder, File, FileFormat};
-use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
-use ssh2_config::{ParseRule, SshConfig};
-use std::env;
-use std::fs::File as StdFile;
-use std::io::{BufReader, ErrorKind};
-use std::path::{Path, PathBuf};
+mod env_defaults;
+mod inventory_loading;
+mod ssh_loading;
 
-// Environment variable names
-/// Environment variable name for controlling error handling behavior.
-///
-/// When set, this variable determines whether Genja should raise (panic/abort) on errors
-/// or handle them gracefully. Accepts boolean-like values such as "true", "false", "yes",
-/// "no", "1", "0", "on", "off" (case-insensitive).
-///
-/// Default: `false` (errors are handled gracefully)
-const ENV_RAISE_ON_ERROR: &str = "GENJA_CORE_RAISE_ON_ERROR";
+#[cfg(test)]
+mod tests;
 
-/// Environment variable name for specifying the inventory plugin.
-///
-/// This variable determines which inventory plugin implementation should be used
-/// for loading and managing host inventory data.
-///
-/// Default: `"FileInventoryPlugin"`
-const ENV_INVENTORY_PLUGIN: &str = "GENJA_INVENTORY_PLUGIN";
+use self::env_defaults::{
+    deserialize_bool_loose, get_default_log_file, get_inventory_plugin_config,
+    get_log_level_default, get_log_to_console_default,
+    get_runner_max_connection_attempts_default, get_runner_max_task_depth_default,
+    get_runner_options_default, get_runner_plugin_default, raise_on_error,
+};
 
-/// Environment variable name for specifying the runner plugin.
-///
-/// This variable determines which runner plugin implementation should be used
-/// for executing tasks across hosts (e.g., "threaded", "serial").
-///
-/// Default: `"threaded"`
-const ENV_RUNNER_PLUGIN: &str = "GENJA_RUNNER_PLUGIN";
-
-/// Environment variable name for setting the logging level.
-///
-/// This variable controls the verbosity of log output. Valid values include
-/// "trace", "debug", "info", "warn", and "error".
-///
-/// Default: `"info"`
-const ENV_LOG_LEVEL: &str = "GENJA_LOGGING_LEVEL";
-
-/// Environment variable name for specifying the log file path.
-///
-/// This variable determines where log output should be written. If not set,
-/// logs are written to `genja.log` in the current working directory.
-///
-/// Default: `genja.log` in the current working directory
-const ENV_LOG_FILE: &str = "GENJA_LOGGING_LOG_FILE";
-
-/// Environment variable name for enabling console logging.
-///
-/// When set, this variable determines whether logs should be written to the console
-/// in addition to the log file. Accepts boolean-like values such as "true", "false",
-/// "yes", "no", "1", "0", "on", "off" (case-insensitive).
-///
-/// Default: `false` (console logging disabled)
-const ENV_LOG_TO_CONSOLE: &str = "GENJA_LOGGING_TO_CONSOLE";
-
-/// Parses a string into a boolean value using loose matching rules.
-///
-/// This function accepts various common string representations of boolean values,
-/// performing case-insensitive matching after trimming whitespace. It recognizes
-/// multiple formats for both true and false values.
-///
-/// # Parameters
-///
-/// * `s` - A string slice containing the value to parse. Leading and trailing
-///   whitespace will be trimmed before parsing.
-///
-/// # Returns
-///
-/// * `Some(true)` - If the string matches any of: "true", "t", "1", "yes", "y", "on"
-///   (case-insensitive)
-/// * `Some(false)` - If the string matches any of: "false", "f", "0", "no", "n", "off"
-///   (case-insensitive)
-/// * `None` - If the string does not match any recognized boolean representation
-fn parse_bool_loose(s: &str) -> Option<bool> {
-    match s.trim().to_ascii_lowercase().as_str() {
-        "true" | "t" | "1" | "yes" | "y" | "on" => Some(true),
-        "false" | "f" | "0" | "no" | "n" | "off" => Some(false),
-        _ => None,
-    }
-}
-
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum BoolLike {
-    Bool(bool),
-    String(String),
-}
-
-fn deserialize_bool_loose<'de, D>(deserializer: D) -> Result<bool, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let value = Option::<BoolLike>::deserialize(deserializer)?;
-    match value {
-        Some(BoolLike::Bool(val)) => Ok(val),
-        Some(BoolLike::String(val)) => parse_bool_loose(val.as_str())
-            .ok_or_else(|| serde::de::Error::custom(format!("invalid boolean value: {val:?}"))),
-        None => Ok(false),
-    }
-}
-
-fn raise_on_error() -> bool {
-    match std::env::var(ENV_RAISE_ON_ERROR) {
-        Ok(s) => match parse_bool_loose(s.as_str()) {
-            Some(true) => true,
-            Some(false) => false,
-            _ => {
-                eprintln!(
-                    "Invalid {} value: {:?}, using default false",
-                    ENV_RAISE_ON_ERROR, s
-                );
-                false
-            }
-        },
-        Err(_) => {
-            eprintln!("{} not found, using default false", ENV_RAISE_ON_ERROR);
-            false
-        }
-    }
-}
-
-/// Retrieves the inventory plugin configuration from environment variables.
-///
-/// This function checks the `GENJA_INVENTORY_PLUGIN` environment variable to determine
-/// which inventory plugin implementation should be used. If the environment variable is
-/// not set or cannot be read, it returns a default value.
-///
-/// # Returns
-///
-/// Returns a `String` containing the name of the inventory plugin to use. If the
-/// `GENJA_INVENTORY_PLUGIN` environment variable is set, returns its value. Otherwise,
-/// returns `"FileInventoryPlugin"` as the default.
-///
-/// See tests in this module for behavioral verification.
-fn get_inventory_plugin_config() -> String {
-    env::var(ENV_INVENTORY_PLUGIN).unwrap_or_else(|_err| String::from("FileInventoryPlugin"))
-}
-
-/// Returns the default runner plugin from `GENJA_RUNNER_PLUGIN`, or "threaded".
-///
-/// See tests in this module for behavioral verification.
-fn get_runner_plugin_default() -> String {
-    env::var(ENV_RUNNER_PLUGIN).unwrap_or_else(|_err| String::from("threaded"))
-}
-
-/// Returns the default runner options JSON.
-///
-/// See tests in this module for behavioral verification.
-fn get_runner_options_default() -> serde_json::Value {
-    serde_json::json!({})
-}
-
-/// Returns the default max task depth for runner execution.
-///
-/// See tests in this module for behavioral verification.
-fn get_runner_max_task_depth_default() -> usize {
-    10
-}
-
-/// Returns the default maximum number of connection attempts for runner execution.
-///
-/// See tests in this module for behavioral verification.
-fn get_runner_max_connection_attempts_default() -> usize {
-    3
-}
-
-/// Returns the default log level from `GENJA_LOGGING_LEVEL`, or "info".
-///
-/// See tests in this module for behavioral verification.
-fn get_log_level_default() -> String {
-    env::var(ENV_LOG_LEVEL).unwrap_or_else(|_err| String::from("info"))
-}
-
-/// Returns the default console logging flag from `GENJA_LOGGING_TO_CONSOLE`.
-///
-/// See tests in this module for behavioral verification.
-fn get_log_to_console_default() -> bool {
-    match env::var(ENV_LOG_TO_CONSOLE) {
-        Ok(val) => parse_bool_loose(val.as_str()).unwrap_or(false),
-        Err(_) => false,
-    }
-}
-
-/// Returns the default log file path, preferring `GENJA_LOGGING_LOG_FILE` when set.
-///
-/// When the environment variable is not set, defaults to `genja.log` in the
-/// current working directory.
-///
-/// See tests in this module for behavioral verification.
-fn get_default_log_file() -> String {
-    match env::var(ENV_LOG_FILE) {
-        Ok(val) => val,
-        Err(_) => {
-            let start_dir = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-            start_dir.join("genja.log").to_string_lossy().to_string()
-        }
-    }
-}
+#[cfg(test)]
+use self::env_defaults::{
+    parse_bool_loose, ENV_INVENTORY_PLUGIN, ENV_LOG_FILE, ENV_LOG_LEVEL, ENV_LOG_TO_CONSOLE,
+    ENV_RAISE_ON_ERROR, ENV_RUNNER_PLUGIN,
+};
 
 /// Configuration options for inventory file paths.
 ///
@@ -586,121 +404,6 @@ impl Default for InventoryConfigBuilder {
     }
 }
 
-impl InventoryConfig {
-    /// Loads inventory data from configured file paths.
-    ///
-    /// This method reads and deserializes inventory files (hosts, groups, and defaults)
-    /// based on the paths specified in the `options` field. If a file path is not
-    /// provided for a particular inventory component, a default or empty value is used.
-    ///
-    /// # Returns
-    ///
-    /// Returns a `Result` containing a tuple of:
-    /// * `Hosts` - The loaded hosts inventory. If no hosts file is specified, returns
-    ///   an empty `Hosts` instance.
-    /// * `Option<Groups>` - The loaded groups inventory, or `None` if no groups file
-    ///   is specified.
-    /// * `Option<Defaults>` - The loaded defaults inventory, or `None` if no defaults
-    ///   file is specified.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`InventoryLoadError`] if:
-    /// * Any specified file cannot be read
-    /// * Any file contains invalid JSON or YAML syntax
-    /// * A file has an unsupported format (not .json, .yaml, or .yml)
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use genja_core::settings::InventoryConfig;
-    ///
-    /// let config = InventoryConfig::default();
-    /// match config.load_inventory_files() {
-    ///     Ok((hosts, groups, defaults)) => {
-    ///         println!("Loaded {} hosts", hosts.len());
-    ///     }
-    ///     Err(e) => eprintln!("Failed to load inventory: {}", e),
-    /// }
-    /// ```
-    pub fn load_inventory_files(
-        &self,
-    ) -> Result<(Hosts, Option<Groups>, Option<Defaults>), InventoryLoadError> {
-        let hosts = match self.options.hosts_file.as_deref() {
-            Some(path) => Self::load_from_file::<Hosts>(InventoryFileKind::Hosts, path)?,
-            None => Hosts::new(),
-        };
-
-        let groups = match self.options.groups_file.as_deref() {
-            Some(path) => Some(Self::load_from_file::<Groups>(InventoryFileKind::Groups, path)?),
-            None => None,
-        };
-
-        let defaults = match self.options.defaults_file.as_deref() {
-            Some(path) => {
-                Some(Self::load_from_file::<Defaults>(InventoryFileKind::Defaults, path)?)
-            }
-            None => None,
-        };
-
-        Ok((hosts, groups, defaults))
-    }
-
-    /// Loads and deserializes data from a file.
-    ///
-    /// This helper method reads a file from the filesystem and deserializes its contents
-    /// based on the file extension. Supports JSON (.json) and YAML (.yaml, .yml) formats.
-    ///
-    /// # Type Parameters
-    ///
-    /// * `T` - The type to deserialize the file contents into. Must implement `DeserializeOwned`.
-    ///
-    /// # Parameters
-    ///
-    /// * `path` - The file path to read and deserialize. The file extension determines
-    ///   the deserialization format.
-    ///
-    /// # Returns
-    ///
-    /// Returns a `Result` containing the deserialized data of type `T`.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`InventoryLoadError`] if:
-    /// * The file cannot be read (e.g., doesn't exist, permission denied)
-    /// * The file contents cannot be parsed as valid JSON or YAML
-    /// * The file has an unsupported extension (not .json, .yaml, or .yml)
-    fn load_from_file<T>(kind: InventoryFileKind, path: &str) -> Result<T, InventoryLoadError>
-    where
-        T: DeserializeOwned,
-    {
-        let contents = std::fs::read_to_string(path).map_err(|e| InventoryLoadError::Read {
-            kind,
-            path: path.to_string(),
-            message: e.to_string(),
-        })?;
-
-        if path.ends_with(".json") {
-            serde_json::from_str(&contents).map_err(|e| InventoryLoadError::ParseJson {
-                kind,
-                path: path.to_string(),
-                message: e.to_string(),
-            })
-        } else if path.ends_with(".yaml") || path.ends_with(".yml") {
-            serde_yaml::from_str(&contents).map_err(|e| InventoryLoadError::ParseYaml {
-                kind,
-                path: path.to_string(),
-                message: e.to_string(),
-            })
-        } else {
-            Err(InventoryLoadError::UnsupportedFormat {
-                kind,
-                path: path.to_string(),
-            })
-        }
-    }
-}
-
 /// Configuration for core Genja behavior.
 ///
 /// This struct controls fundamental aspects of how Genja operates, particularly
@@ -857,199 +560,6 @@ impl Default for CoreConfigBuilder {
 pub struct SSHConfig {
     config_file: Option<String>,
 }
-impl SSHConfig {
-    /// Validates the SSH configuration file syntax if a path is provided.
-    ///
-    /// This method performs comprehensive validation of an SSH configuration file by:
-    /// 1. Verifying that the file exists and is accessible
-    /// 2. Opening the file for reading
-    /// 3. Parsing the file contents using strict SSH config syntax rules
-    ///
-    /// If no SSH configuration file is specified (the `config_file` field is `None`),
-    /// this method returns `Ok(())` without performing any validation.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if:
-    /// * No config file is specified (nothing to validate)
-    /// * The config file exists, can be opened, and contains valid SSH configuration syntax
-    ///
-    /// Returns `Err(SshConfigError)` if:
-    /// * The specified file does not exist or cannot be accessed
-    /// * The file cannot be opened due to permission issues or other I/O errors
-    /// * The file contents cannot be parsed as valid SSH configuration syntax
-    ///
-    /// # Errors
-    ///
-    /// This method returns an error in the following cases:
-    /// * File existence check fails (see `ensure_exists` for details)
-    /// * `SshConfigError::OpenFailed` - The file exists but cannot be opened
-    /// * `SshConfigError::ParseFailed` - The file contains invalid SSH config syntax
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use genja_core::settings::SSHConfig;
-    ///
-    /// let config = SSHConfig::builder()
-    ///     .config_file("/home/user/.ssh/config")
-    ///     .build();
-    ///
-    /// match config.validate() {
-    ///     Ok(()) => println!("SSH config is valid"),
-    ///     Err(e) => eprintln!("Invalid SSH config: {}", e),
-    /// }
-    /// ```
-    pub fn validate(&self) -> Result<(), SshConfigError> {
-        if let Some(ref path) = self.config_file {
-            let path = Path::new(path);
-
-            self.ensure_exists(path)?;
-
-            let file = match StdFile::open(path) {
-                Ok(file) => file,
-                Err(e) => {
-                    return Err(SshConfigError::OpenFailed {
-                        path: path.display().to_string(),
-                        message: e.to_string(),
-                    })
-                }
-            };
-            let mut reader = BufReader::new(file);
-
-            match SshConfig::default().parse(&mut reader, ParseRule::STRICT) {
-                Ok(_) => Ok(()),
-                Err(e) => Err(SshConfigError::ParseFailed {
-                    path: path.display().to_string(),
-                    message: e.to_string(),
-                }),
-            }
-        } else {
-            Ok(()) // No config file specified, nothing to validate
-        }
-    }
-
-    /// Parses the SSH configuration file and returns the parsed configuration.
-    ///
-    /// This method reads and parses an SSH configuration file if one is specified in the
-    /// `config_file` field. The parsing follows strict SSH config file syntax rules as
-    /// defined by OpenSSH. If no configuration file is specified, the method returns
-    /// `Ok(None)` without performing any parsing.
-    ///
-    /// # Returns
-    ///
-    /// Returns a `Result` containing:
-    /// * `Ok(Some(SshConfig))` - If a config file is specified and successfully parsed,
-    ///   containing the parsed SSH configuration with all host entries and settings.
-    /// * `Ok(None)` - If no config file is specified (the `config_file` field is `None`).
-    /// * `Err(SshConfigError)` - If an error occurs during parsing.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`SshConfigError`] if:
-    /// * The specified SSH config file does not exist at the given path
-    /// * The file cannot be opened due to permission issues or other I/O errors
-    /// * The file contents cannot be parsed as valid SSH configuration syntax
-    /// * The file contains syntax errors or invalid SSH configuration directives
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use genja_core::settings::SSHConfig;
-    ///
-    /// let config = SSHConfig::builder()
-    ///     .config_file("/home/user/.ssh/config")
-    ///     .build();
-    ///
-    /// match config.parse() {
-    ///     Ok(Some(ssh_config)) => {
-    ///         println!("Successfully parsed SSH config");
-    ///     }
-    ///     Ok(None) => {
-    ///         println!("No SSH config file specified");
-    ///     }
-    ///     Err(e) => {
-    ///         eprintln!("Failed to parse SSH config: {}", e);
-    ///     }
-    /// }
-    /// ```
-    pub fn parse(&self) -> Result<Option<SshConfig>, SshConfigError> {
-        if let Some(ref path) = self.config_file {
-            let path = Path::new(path);
-
-            self.ensure_exists(path)?;
-
-            let file = match StdFile::open(path) {
-                Ok(file) => file,
-                Err(e) => Err(SshConfigError::OpenFailed {
-                    path: path.display().to_string(),
-                    message: e.to_string(),
-                })?,
-            };
-            let mut reader = BufReader::new(file);
-
-            match SshConfig::default().parse(&mut reader, ParseRule::STRICT) {
-                Ok(config) => Ok(Some(config)),
-                Err(e) => Err(SshConfigError::ParseFailed {
-                    path: path.display().to_string(),
-                    message: e.to_string(),
-                }),
-            }
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Verifies that an SSH configuration file exists and is accessible.
-    ///
-    /// This method checks whether the specified file path exists and can be accessed.
-    /// It provides detailed error messages for different failure scenarios, including
-    /// permission issues and I/O errors.
-    ///
-    /// # Parameters
-    ///
-    /// * `path` - A reference to the file path to check. This should point to an SSH
-    ///   configuration file that needs to be validated for existence and accessibility.
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if the file exists and is accessible. Returns `Err(SshConfigError)` if
-    /// a descriptive error message if:
-    /// * The file does not exist
-    /// * Permission is denied when attempting to access the file
-    /// * An I/O error occurs during the existence check
-    /// * Any other filesystem error prevents verification
-    ///
-    /// # Errors
-    ///
-    /// This method returns an error in the following cases:
-    /// * `SshConfigError::NotFound` - The file does not exist at the specified path
-    /// * `SshConfigError::PermissionDenied` - The file exists but cannot be accessed due to
-    ///   insufficient permissions
-    /// * `SshConfigError::CheckFailed` - Any other filesystem error occurred during the check
-    fn ensure_exists(&self, path: &Path) -> Result<(), SshConfigError> {
-        match path.try_exists() {
-            Ok(true) => Ok(()),
-            Ok(false) => Err(SshConfigError::NotFound {
-                path: path.display().to_string(),
-            }),
-            Err(e) => match e.kind() {
-                ErrorKind::PermissionDenied => Err(SshConfigError::PermissionDenied {
-                    path: path.display().to_string(),
-                    message: e.to_string(),
-                }),
-                ErrorKind::NotFound => Err(SshConfigError::NotFound {
-                    path: path.display().to_string(),
-                }),
-                _ => Err(SshConfigError::CheckFailed {
-                    path: path.display().to_string(),
-                    message: e.to_string(),
-                }),
-            },
-        }
-    }
-}
-
 impl Default for SSHConfig {
     fn default() -> Self {
         SSHConfig { config_file: None }
@@ -1803,608 +1313,5 @@ impl Default for SettingsBuilder {
             runner: None,
             logging: None,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{OptionsConfig, RunnerConfig, SSHConfig};
-    use regex::Regex;
-    use serde_json::json;
-    use std::env;
-    use std::io::Write;
-    use std::path::{Path, PathBuf};
-    use std::sync::{Mutex, OnceLock};
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    /// Returns a static reference to a mutex used for synchronizing environment variable access in tests.
-    ///
-    /// This function ensures that tests modifying environment variables do not run concurrently,
-    /// preventing race conditions and test interference. The mutex is initialized once and reused
-    /// across all test invocations.
-    ///
-    /// # Returns
-    ///
-    /// A static reference to a `Mutex<()>` that can be locked to serialize environment variable operations.
-    fn env_lock() -> &'static Mutex<()> {
-        static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-        LOCK.get_or_init(|| Mutex::new(()))
-    }
-
-    /// Temporarily sets or removes an environment variable for the duration of a test function.
-    ///
-    /// This function provides a safe way to test code that depends on environment variables by:
-    /// 1. Acquiring an exclusive lock to prevent concurrent environment modifications
-    /// 2. Saving the current value of the environment variable
-    /// 3. Setting or removing the environment variable as specified
-    /// 4. Executing the provided test function
-    /// 5. Restoring the original environment variable state
-    ///
-    /// # Parameters
-    ///
-    /// * `key` - The name of the environment variable to modify
-    /// * `val` - The value to set for the environment variable. If `Some(value)`, the variable
-    ///   is set to that value. If `None`, the variable is removed from the environment.
-    /// * `f` - A closure containing the test code to execute with the modified environment variable
-    fn with_env_var(key: &str, val: Option<&str>, f: impl FnOnce()) {
-        let _guard = env_lock().lock().unwrap();
-        let prev = env::var(key).ok();
-        match val {
-            Some(v) => env::set_var(key, v),
-            None => env::remove_var(key), // tests when the variable is not set
-        }
-        f();
-        match prev {
-            Some(v) => env::set_var(key, v),
-            None => env::remove_var(key),
-        }
-    }
-
-    struct Context {
-        _tempdir: tempfile::TempDir,
-        filename: PathBuf,
-    }
-
-    fn write_temp_ssh_config(contents: &str) -> Context {
-        let tempdir = tempfile::tempdir().unwrap();
-        let unique = format!(
-            "sshconfig_test_{}_{}",
-            std::process::id(),
-            SystemTime::now()
-                .duration_since(UNIX_EPOCH)
-                .unwrap()
-                .as_nanos()
-        );
-        let filename = tempdir.path().join(unique);
-        let mut file = std::fs::File::create(&filename).unwrap();
-        file.write_all(contents.as_bytes()).unwrap();
-        Context {
-            _tempdir: tempdir,
-            filename,
-        }
-    }
-
-    #[test]
-    fn validate_ok_with_valid_config() {
-        let context = write_temp_ssh_config("Host example\n  HostName example.com\n");
-        let ssh_config = SSHConfig {
-            config_file: Some(context.filename.to_string_lossy().to_string()),
-        };
-
-        let result = ssh_config.validate();
-        assert!(result.is_ok());
-        assert!(matches!(result, Ok(_)));
-    }
-
-    #[test]
-    fn validate_ok_with_no_config_file() {
-        let ssh_config = SSHConfig { config_file: None };
-        assert!(ssh_config.validate().is_ok());
-    }
-
-    #[test]
-    fn validate_err_with_invalid_config() {
-        let context = write_temp_ssh_config("Contents that are not valid ssh config contents\n");
-        let ssh_config = SSHConfig {
-            config_file: Some(context.filename.to_string_lossy().to_string()),
-        };
-        let result = ssh_config.validate();
-        assert!(matches!(
-            result,
-            Err(crate::SshConfigError::ParseFailed { .. })
-        ));
-        let pattern =
-            Regex::new(r"Failed to parse SSH config file \S+: unknown field: Contents").unwrap();
-        assert!(pattern.is_match(&result.unwrap_err().to_string()));
-    }
-
-    #[test]
-    fn parse_returns_config_when_present() {
-        let context = write_temp_ssh_config("Host example\n  HostName example.com\n");
-        let ssh_config = SSHConfig {
-            config_file: Some(context.filename.to_string_lossy().to_string()),
-        };
-
-        let result = ssh_config.parse();
-        assert!(matches!(result, Ok(Some(_))));
-    }
-
-    #[test]
-    fn parse_returns_none_when_missing() {
-        let ssh_config = SSHConfig { config_file: None };
-        assert!(matches!(ssh_config.parse(), Ok(None)));
-    }
-
-    #[test]
-    fn ensure_exists_returns_ok_when_present() {
-        let context = write_temp_ssh_config("Host example\n  HostName example.com\n");
-        let ssh_config = SSHConfig {
-            config_file: Some(context.filename.to_string_lossy().to_string()),
-        };
-
-        let result = ssh_config.ensure_exists(&context.filename);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn ensure_exists_returns_err_when_missing() {
-        let ssh_config = SSHConfig { config_file: None };
-        let result = ssh_config.ensure_exists(&Path::new("nonexistent_file.txt"));
-        assert!(matches!(
-            result,
-            Err(crate::SshConfigError::NotFound { .. })
-        ));
-        assert_eq!(
-            result.unwrap_err().to_string(),
-            "SSH config file not found: nonexistent_file.txt"
-        );
-    }
-
-    #[test]
-    fn options_config_default_is_all_none() {
-        let options = OptionsConfig::default();
-        assert!(options.hosts_file.is_none());
-        assert!(options.groups_file.is_none());
-        assert!(options.defaults_file.is_none());
-    }
-
-    #[test]
-    fn options_config_deserializes_empty_object_to_none() {
-        let options: OptionsConfig = serde_json::from_str("{}").unwrap();
-        assert!(options.hosts_file.is_none());
-        assert!(options.groups_file.is_none());
-        assert!(options.defaults_file.is_none());
-    }
-
-    #[test]
-    fn options_config_deserializes_with_values() {
-        let json = r#"{
-            "hosts_file": "/tmp/hosts.yaml",
-            "groups_file": "/tmp/groups.yaml",
-            "defaults_file": "/tmp/defaults.yaml"
-        }"#;
-        let options: OptionsConfig = serde_json::from_str(json).unwrap();
-        assert_eq!(options.hosts_file.as_deref(), Some("/tmp/hosts.yaml"));
-        assert_eq!(options.groups_file.as_deref(), Some("/tmp/groups.yaml"));
-        assert_eq!(options.defaults_file.as_deref(), Some("/tmp/defaults.yaml"));
-    }
-
-    #[test]
-    fn runner_config_default_values() {
-        let runner = RunnerConfig::default();
-        assert_eq!(runner.plugin, "threaded");
-        assert_eq!(runner.options, json!({}));
-        assert_eq!(runner.worker_count, None);
-        assert_eq!(runner.max_task_depth, 10);
-        assert_eq!(runner.max_connection_attempts, 3);
-    }
-
-    #[test]
-    fn runner_config_deserializes_empty_object_to_defaults() {
-        let runner: RunnerConfig = serde_json::from_str("{}").unwrap();
-        assert_eq!(runner.plugin, "threaded");
-        assert_eq!(runner.options, json!({}));
-        assert_eq!(runner.worker_count, None);
-        assert_eq!(runner.max_task_depth, 10);
-        assert_eq!(runner.max_connection_attempts, 3);
-    }
-
-    #[test]
-    fn runner_config_deserializes_with_values() {
-        let json = r#"{
-            "plugin": "custom",
-            "options": {"queue": "fast", "strategy": "burst"},
-            "worker_count": 6,
-            "max_task_depth": 5,
-            "max_connection_attempts": 7
-        }"#;
-        let runner: RunnerConfig = serde_json::from_str(json).unwrap();
-        assert_eq!(runner.plugin, "custom");
-        assert_eq!(
-            runner.options,
-            json!({"queue": "fast", "strategy": "burst"})
-        );
-        assert_eq!(runner.worker_count, Some(6));
-        assert_eq!(runner.max_task_depth, 5);
-        assert_eq!(runner.max_connection_attempts, 7);
-    }
-
-    #[test]
-    fn runner_config_builder_sets_max_connection_attempts() {
-        let runner = RunnerConfig::builder().max_connection_attempts(9).build();
-
-        assert_eq!(runner.max_connection_attempts(), 9);
-    }
-
-    #[test]
-    fn runner_config_builder_sets_worker_count() {
-        let runner = RunnerConfig::builder().worker_count(4).build();
-
-        assert_eq!(runner.worker_count(), Some(4));
-    }
-
-    #[test]
-    fn parse_bool_loose_accepts_common_values() {
-        assert_eq!(super::parse_bool_loose("true"), Some(true));
-        assert_eq!(super::parse_bool_loose("TrUe"), Some(true));
-        assert_eq!(super::parse_bool_loose("1"), Some(true));
-        assert_eq!(super::parse_bool_loose("yes"), Some(true));
-        assert_eq!(super::parse_bool_loose("on"), Some(true));
-        assert_eq!(super::parse_bool_loose("false"), Some(false));
-        assert_eq!(super::parse_bool_loose("0"), Some(false));
-        assert_eq!(super::parse_bool_loose("no"), Some(false));
-        assert_eq!(super::parse_bool_loose("off"), Some(false));
-        assert_eq!(super::parse_bool_loose("maybe"), None);
-    }
-
-    #[test]
-    fn deserialize_bool_loose_from_string_and_bool() {
-        #[derive(serde::Deserialize)]
-        struct T {
-            #[serde(deserialize_with = "super::deserialize_bool_loose")]
-            v: bool,
-        }
-
-        let t: T = serde_json::from_str(r#"{ "v": "yes" }"#).unwrap();
-        assert!(t.v);
-        let t: T = serde_json::from_str(r#"{ "v": false }"#).unwrap();
-        assert!(!t.v);
-    }
-
-    #[test]
-    fn deserialize_bool_loose_rejects_invalid_string() {
-        #[derive(serde::Deserialize, Debug)]
-        struct T {
-            #[serde(deserialize_with = "super::deserialize_bool_loose")]
-            _v: bool,
-        }
-
-        let err = serde_json::from_str::<T>(r#"{ "_v": "maybe" }"#).unwrap_err();
-        assert!(err.to_string().contains("invalid boolean value"));
-    }
-
-    #[test]
-    fn raise_on_error_uses_env_and_fallbacks() {
-        with_env_var(super::ENV_RAISE_ON_ERROR, Some("true"), || {
-            assert!(super::raise_on_error());
-        });
-        with_env_var(super::ENV_RAISE_ON_ERROR, Some("not_a_bool"), || {
-            assert!(!super::raise_on_error());
-        });
-        with_env_var(super::ENV_RAISE_ON_ERROR, None, || {
-            assert!(!super::raise_on_error());
-        });
-    }
-
-    #[test]
-    fn get_log_to_console_default_parses_env() {
-        with_env_var(super::ENV_LOG_TO_CONSOLE, Some("yes"), || {
-            assert!(super::get_log_to_console_default());
-        });
-        with_env_var(super::ENV_LOG_TO_CONSOLE, Some("no"), || {
-            assert!(!super::get_log_to_console_default());
-        });
-    }
-
-    #[test]
-    fn env_string_defaults_respect_env_and_fallbacks() {
-        with_env_var(super::ENV_INVENTORY_PLUGIN, Some("CustomInv"), || {
-            assert_eq!(super::get_inventory_plugin_config(), "CustomInv");
-        });
-        with_env_var(super::ENV_INVENTORY_PLUGIN, None, || {
-            assert_eq!(super::get_inventory_plugin_config(), "FileInventoryPlugin");
-        });
-
-        with_env_var(super::ENV_RUNNER_PLUGIN, Some("CustomRunner"), || {
-            assert_eq!(super::get_runner_plugin_default(), "CustomRunner");
-        });
-        with_env_var(super::ENV_RUNNER_PLUGIN, None, || {
-            assert_eq!(super::get_runner_plugin_default(), "threaded");
-        });
-
-        with_env_var(super::ENV_LOG_LEVEL, Some("debug"), || {
-            assert_eq!(super::get_log_level_default(), "debug");
-        });
-        with_env_var(super::ENV_LOG_LEVEL, None, || {
-            assert_eq!(super::get_log_level_default(), "info");
-        });
-    }
-
-    #[test]
-    fn get_default_log_file_prefers_env() {
-        with_env_var(super::ENV_LOG_FILE, Some("/tmp/genja-test.log"), || {
-            assert_eq!(super::get_default_log_file(), "/tmp/genja-test.log");
-        });
-    }
-
-    #[test]
-    fn get_default_log_file_uses_cwd_when_env_missing() {
-        let _guard = env_lock().lock().unwrap();
-        let prev = env::var(super::ENV_LOG_FILE).ok();
-        env::remove_var(super::ENV_LOG_FILE);
-
-        let tempdir = tempfile::tempdir().unwrap();
-        let prev_dir = env::current_dir().unwrap();
-        env::set_current_dir(tempdir.path()).unwrap();
-
-        let expected = tempdir.path().join("genja.log");
-        assert_eq!(
-            super::get_default_log_file(),
-            expected.to_string_lossy().to_string()
-        );
-
-        env::set_current_dir(prev_dir).unwrap();
-        match prev {
-            Some(v) => env::set_var(super::ENV_LOG_FILE, v),
-            None => env::remove_var(super::ENV_LOG_FILE),
-        }
-    }
-
-    #[test]
-    fn settings_from_file_errors_when_missing() {
-        let tempdir = tempfile::tempdir().unwrap();
-        let missing = tempdir.path().join("missing.yaml");
-        let err = super::Settings::from_file(missing.to_string_lossy().as_ref()).unwrap_err();
-        assert!(err.to_string().to_lowercase().contains("not found"));
-    }
-
-    #[test]
-    fn settings_from_file_errors_on_unsupported_extension() {
-        let tempdir = tempfile::tempdir().unwrap();
-        let file_path = tempdir.path().join("config.txt");
-        std::fs::write(&file_path, "{}").unwrap();
-        let err = super::Settings::from_file(file_path.to_string_lossy().as_ref()).unwrap_err();
-        assert!(matches!(
-            err,
-            crate::ConfigLoadError::UnsupportedFormat { .. }
-        ));
-    }
-
-    #[test]
-    fn settings_from_file_errors_on_invalid_json() {
-        let tempdir = tempfile::tempdir().unwrap();
-        let file_path = tempdir.path().join("config.json");
-        std::fs::write(&file_path, "{ not valid json").unwrap();
-        let err = super::Settings::from_file(file_path.to_string_lossy().as_ref()).unwrap_err();
-        assert!(!err.to_string().is_empty());
-    }
-
-    #[test]
-    fn settings_from_file_uses_defaults_for_empty_file() {
-        let _guard = env_lock().lock().unwrap();
-        let keys = [
-            super::ENV_RAISE_ON_ERROR,
-            super::ENV_INVENTORY_PLUGIN,
-            super::ENV_RUNNER_PLUGIN,
-            super::ENV_LOG_LEVEL,
-            super::ENV_LOG_FILE,
-            super::ENV_LOG_TO_CONSOLE,
-        ];
-
-        let prev: Vec<(String, Option<String>)> = keys
-            .iter()
-            .map(|key| (key.to_string(), env::var(key).ok()))
-            .collect();
-
-        for key in keys {
-            env::remove_var(key);
-        }
-
-        let tempdir = tempfile::tempdir().unwrap();
-        let file_path = tempdir.path().join("config.yaml");
-        std::fs::write(&file_path, "{}").unwrap();
-        let settings = super::Settings::from_file(file_path.to_string_lossy().as_ref()).unwrap();
-
-        assert!(!settings.core().raise_on_error());
-        assert_eq!(settings.inventory().plugin(), "FileInventoryPlugin");
-        assert_eq!(settings.runner().plugin(), "threaded");
-        assert_eq!(settings.logging().level(), "info");
-        assert!(settings.logging().enabled());
-
-        for (key, val) in prev {
-            match val {
-                Some(v) => env::set_var(&key, v),
-                None => env::remove_var(&key),
-            }
-        }
-    }
-
-    #[test]
-    fn settings_from_file_loads_populated_yaml() {
-        let ssh_context = write_temp_ssh_config("Host example\n  HostName example.com\n");
-        let tempdir = tempfile::tempdir().unwrap();
-        let file_path = tempdir.path().join("config.yaml");
-        let yaml = format!(
-            r#"
-core:
-  raise_on_error: "yes"
-inventory:
-  plugin: "CustomInventoryPlugin"
-  options:
-    hosts_file: "./inventory/hosts.yaml"
-    groups_file: "./inventory/groups.yaml"
-    defaults_file: "./inventory/defaults.yaml"
-  transform_function: "normalize_inventory"
-  transform_function_options:
-    mode: "strict"
-    retries: 2
-ssh:
-  config_file: "{}"
-runner:
-  plugin: "serial"
-  options:
-    queue: "fast"
-  worker_count: 6
-  max_task_depth: 5
-  max_connection_attempts: 7
-logging:
-  enabled: "no"
-  level: "debug"
-  log_file: "./custom.log"
-  to_console: "yes"
-  file_size: 2048
-  max_file_count: 4
-"#,
-            ssh_context.filename.display()
-        );
-        std::fs::write(&file_path, yaml).unwrap();
-
-        let settings = super::Settings::from_file(file_path.to_string_lossy().as_ref()).unwrap();
-
-        assert!(settings.core().raise_on_error());
-        assert_eq!(settings.inventory().plugin(), "CustomInventoryPlugin");
-        assert_eq!(
-            settings.inventory().options().hosts_file(),
-            Some("./inventory/hosts.yaml")
-        );
-        assert_eq!(
-            settings.inventory().options().groups_file(),
-            Some("./inventory/groups.yaml")
-        );
-        assert_eq!(
-            settings.inventory().options().defaults_file(),
-            Some("./inventory/defaults.yaml")
-        );
-        assert_eq!(
-            settings.inventory().transform_function(),
-            Some("normalize_inventory")
-        );
-        assert_eq!(
-            settings
-                .inventory()
-                .transform_function_options()
-                .and_then(|options| options.get("mode"))
-                .and_then(|value| value.as_str()),
-            Some("strict")
-        );
-        assert_eq!(
-            settings
-                .inventory()
-                .transform_function_options()
-                .and_then(|options| options.get("retries"))
-                .and_then(|value| value.as_u64()),
-            Some(2)
-        );
-        assert_eq!(
-            settings.ssh().config_file(),
-            Some(ssh_context.filename.to_string_lossy().as_ref())
-        );
-        assert_eq!(settings.runner().plugin(), "serial");
-        assert_eq!(settings.runner().options(), &json!({ "queue": "fast" }));
-        assert_eq!(settings.runner().worker_count(), Some(6));
-        assert_eq!(settings.runner().max_task_depth(), 5);
-        assert_eq!(settings.runner().max_connection_attempts(), 7);
-        assert!(!settings.logging().enabled());
-        assert_eq!(settings.logging().level(), "debug");
-        assert_eq!(settings.logging().log_file(), "./custom.log");
-        assert!(settings.logging().to_console());
-        assert_eq!(settings.logging().file_size(), 2048);
-        assert_eq!(settings.logging().max_file_count(), 4);
-    }
-
-    #[test]
-    fn settings_from_file_errors_on_invalid_ssh_config() {
-        let ssh_context = write_temp_ssh_config("Contents that are not valid ssh config contents\n");
-        let tempdir = tempfile::tempdir().unwrap();
-        let file_path = tempdir.path().join("config.yaml");
-        let yaml = format!(
-            r#"
-ssh:
-  config_file: "{}"
-"#,
-            ssh_context.filename.display()
-        );
-        std::fs::write(&file_path, yaml).unwrap();
-
-        let err = super::Settings::from_file(file_path.to_string_lossy().as_ref()).unwrap_err();
-        assert!(matches!(
-            err,
-            crate::ConfigLoadError::SshConfig(crate::SshConfigError::ParseFailed { .. })
-        ));
-    }
-
-    #[test]
-    fn inventory_loads_empty_files() {
-        let tempdir = tempfile::tempdir().unwrap();
-        let hosts_path = tempdir.path().join("hosts.json");
-        let groups_path = tempdir.path().join("groups.json");
-        let defaults_path = tempdir.path().join("defaults.json");
-        std::fs::write(&hosts_path, "{}").unwrap();
-        std::fs::write(&groups_path, "{}").unwrap();
-        std::fs::write(&defaults_path, "{}").unwrap();
-
-        let options = super::OptionsConfig::builder()
-            .hosts_file(hosts_path.to_string_lossy().as_ref())
-            .groups_file(groups_path.to_string_lossy().as_ref())
-            .defaults_file(defaults_path.to_string_lossy().as_ref())
-            .build();
-        let config = super::InventoryConfig::builder().options(options).build();
-
-        let (hosts, groups, defaults) = config.load_inventory_files().unwrap();
-        assert!(hosts.is_empty());
-        assert!(groups.unwrap().is_empty());
-        // let defaults = defaults.unwrap();
-        assert!(defaults.unwrap().is_empty());
-        // assert!(defaults
-        //     .as_object()
-        //     .map(|map| map.is_empty())
-        //     .unwrap_or(false));
-    }
-
-    #[test]
-    fn inventory_load_errors_on_missing_file() {
-        let tempdir = tempfile::tempdir().unwrap();
-        let missing = tempdir.path().join("missing.json");
-        let options = super::OptionsConfig::builder()
-            .hosts_file(missing.to_string_lossy().as_ref())
-            .build();
-        let config = super::InventoryConfig::builder().options(options).build();
-        let err = config.load_inventory_files().unwrap_err();
-        assert!(matches!(
-            err,
-            crate::InventoryLoadError::Read {
-                kind: crate::InventoryFileKind::Hosts,
-                ..
-            }
-        ));
-    }
-
-    #[test]
-    fn inventory_load_errors_on_unsupported_extension() {
-        let tempdir = tempfile::tempdir().unwrap();
-        let file_path = tempdir.path().join("hosts.txt");
-        std::fs::write(&file_path, "{}").unwrap();
-        let options = super::OptionsConfig::builder()
-            .hosts_file(file_path.to_string_lossy().as_ref())
-            .build();
-        let config = super::InventoryConfig::builder().options(options).build();
-        let err = config.load_inventory_files().unwrap_err();
-        assert!(matches!(
-            err,
-            crate::InventoryLoadError::UnsupportedFormat {
-                kind: crate::InventoryFileKind::Hosts,
-                ..
-            }
-        ));
     }
 }
